refactor: merge profile view model
  * merge UserProfileViewModel.kt into ProfileViewModel.kt

  * update imports in Profile.kt
